refactor existing e2e test for reuse
