+++ openshift-with-appstudio-test/e2e/const.go
@@ -1,0 +1,8 @@
+package e2e
+
+const (
+	testNamespace          = "jvm-build-service-test-namespace-"
+	maxNameLength          = 63
+	randomLength           = 5
+	maxGeneratedNameLength = maxNameLength - randomLength
+)
